Various FoundBugs repaired

+++ groove/trunk/src/groove/gui/jgraph/EditorJGraph.java
@@ -115,8 +115,8 @@ public class EditorJGraph extends JGraph {
         // set the bounds and store them in the cell
         Dimension size = JAttr.DEFAULT_NODE_SIZE;
         Point2D corner =
-            new Double(atPoint.getX() - size.width / 2, atPoint.getY()
-                - size.height / 2);
+            new Double(atPoint.getX() - (double) size.width / 2, atPoint.getY()
+                - (double) size.height / 2);
         GraphConstants.setBounds(
             jVertex.getAttributes(),
             new SerializableRectangle2D(corner.getX(), corner.getY(),
